Fix warnings

+++ tools/toktx/image.hpp
@@ -199,9 +199,9 @@ struct vec3 : public vec3_base {
         if (gc_m[i] == r && gc_m[i] == g && gc_m[i] == b) {
             return;
         } else {
-            r = r / gc_s[i] * 2.0f - 1.0f;
-            g = g / gc_s[i] * 2.0f - 1.0f;
-            b = b / gc_s[i] * 2.0f - 1.0f;
+            r = (r / (double)gc_s[i]) * 2.0f - 1.0f;
+            g = (g / (double)gc_s[i]) * 2.0f - 1.0f;
+            b = (b / (double)gc_s[i]) * 2.0f - 1.0f;
             clamp(-1.0f, 1.0f);
             base_normalize();
             r = (std::floor((r + 1.0f) * gc_s[i] * 0.5f + 0.5f));
